KS-1-4 Иземенил тип переменной на Double

+++ src/main/kotlin/lesson_1/lesson_1_task_4.kt
@@ -5,6 +5,6 @@ fun main() {
     var age: Byte = 27
     var partOfDay: Float = 0.075F
     var seconds: Short = 6480
-    var partOfYear: Float = 2.0547945205479453E-4F
+    var partOfYear: Double = 2.0547945205479453E-4
     var apogee: Int = 327000
 }
